requirements.txt
transformers
torch
torchvision
gradio
Pillow
pandas
insightface
scikit-learn
pytesseract 